Refactoring. Merger inn endringer i vedlegg OpplysningerBruktIBeregningUT.
