+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/maler/fraser/Omsorgsopptjening.kt
@@ -3,6 +3,7 @@ package no.nav.pensjon.brev.maler.fraser
 import no.nav.pensjon.brev.api.model.*
 import no.nav.pensjon.brev.api.model.FellesSelectors.avsenderEnhet
 import no.nav.pensjon.brev.api.model.NAVEnhetSelectors.nettside
+import no.nav.pensjon.brev.maler.fraser.common.Constants
 import no.nav.pensjon.brev.model.format
 import no.nav.pensjon.brev.template.*
 import no.nav.pensjon.brev.template.Language.*
