+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/maler/fraser/common/Vedtak.kt
@@ -1,4 +1,4 @@
-package no.nav.pensjon.brev.maler.fraser.vedtak
+package no.nav.pensjon.brev.maler.fraser.common
 
 import no.nav.pensjon.brev.template.*
 import no.nav.pensjon.brev.template.Language.*
